change spring annotation

+++ src/main/java/com/iemdb/iemdb/Repository/CommentRepository.java
@@ -3,5 +3,5 @@ package com.iemdb.iemdb.Repository;
 import com.iemdb.iemdb.Model.Comment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends CrudRepository<Comment, Integer> {
 }
